+++ wagon_common/helpers/gh/url.py
@@ -2,9 +2,8 @@
 import re
 import os
 
-from colorama import Fore, Style
-
 from wagon_common.helpers.git.clone import clone_repo
+from wagon_common.helpers.output import red
 
 
 class GitHubRepo:
@@ -47,11 +46,9 @@ class GitHubRepo:
 
         if os.path.isdir(path):
 
-            print(Fore.RED
-                  + "\nDestination directory already exists 🤒"
-                  + Style.RESET_ALL
-                  + "\nCannot clone the repo to existing location"
-                  + f"\n- path: {path}")
+            red("\nDestination directory already exists 🤒",
+                "\nCannot clone the repo to existing location"
+                + f"\n- path: {path}")
 
             raise FileExistsError(f"Repo directory already exists: {path}")
 
@@ -59,14 +56,12 @@ class GitHubRepo:
 
         if not cloned:
 
-            print(Fore.RED
-                  + "\nUnable to clone repo 🤒"
-                  + Style.RESET_ALL
-                  + "\nCannot clone the repo to existing location"
-                  + f"\n- url: {self.url}"
-                  + f"\n- path: {path}"
-                  + f"\n- output: {output}"
-                  + f"\n- error: {error}")
+            red("\nUnable to clone repo 🤒",
+                "\nCannot clone the repo to existing location"
+                + f"\n- url: {self.url}"
+                + f"\n- path: {path}"
+                + f"\n- output: {output}"
+                + f"\n- error: {error}")
 
             raise Exception(f"Unable to clone repo: {self.url}, {path}")