+++ wagon_common/helpers/scope.py
@@ -8,7 +8,7 @@ from wagon_common.helpers.output import print_files
 import os
 import glob
 
-from colorama import Fore, Style
+from wagon_common.helpers.output import blue
 
 
 def resolve_scope(sources, patterns, return_inexisting=False, verbose=False):
@@ -36,9 +36,7 @@ def resolve_scope(sources, patterns, return_inexisting=False, verbose=False):
     for pattern in patterns:
 
         if verbose:
-            print(Fore.BLUE
-                  + f"\nResolving files matching the {pattern} pattern in the scope..."
-                  + Style.RESET_ALL)
+            blue(f"\nResolving files matching the {pattern} pattern in the scope...")
 
         # remove all sources that do not match the pattern
         filtered_sources = []