+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 
+# 0.2.9 (2022-11-25)
+
+### Added
+
+- Adds `manage_command` command subprocess helper handling output for `run_command`
+- Adds `black`, `red`, `green`, `yellow`, `blue`, `magenta`, `cyan` and `white` helpers for print with `colorama`
+
+### Updated
+
+- Update `run_command` to print command output progressively (used in order to display remote command output when calling `gcloud compute ssh ...`)
+
 # 0.2.8 (2022-11-25)
 
 ### Updated